consumer/requirements.txt
kafka-python
boto3
pydantic